refactor: rewrite using dashboard SDK
