+++ developer-assist-dashboard/package.json
@@ -6,22 +6,18 @@
   },
   "private": true,
   "dependencies": {
-    "@fluentui/react": "^8.107.1",
     "@fluentui/react-components": "^9.18.0",
     "@fluentui/react-icons": "^2.0.186",
-    "@fluentui/react-icons-mdl2": "^1.3.30",
     "@microsoft/microsoft-graph-client": "^3.0.5",
     "@microsoft/teams-js": "^2.7.1",
     "@microsoft/teamsfx": "^2.2.0",
     "@microsoft/teamsfx-react": "^3.0.0-beta",
     "azure-devops-node-api": "^12.0.0",
-    "node-fetch": "^2.6.7",
     "octokit": "^2.0.14",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.0",
-    "react-scripts": "^5.0.1",
-    "react-syntax-highlighter": "^15.5.0"
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
     "@types/node": "^14.0.0",
@@ -29,7 +25,6 @@
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@types/react-router-dom": "^5.3.3",
-    "@types/react-syntax-highlighter": "^15.5.6",
     "cross-env": "^7.0.3",
     "env-cmd": "^10.1.0",
     "typescript": "^4.1.2"
