+++ developer-assist-dashboard/src/App.tsx
@@ -1,36 +1,39 @@
 import "./App.css";
 
-import {
-  HashRouter as Router,
-  Navigate,
-  Route,
-  Routes,
-} from "react-router-dom";
+import { useEffect } from "react";
+import { HashRouter as Router, Navigate, Route, Routes } from "react-router-dom";
 
-// https://fluentsite.z22.web.core.windows.net/quick-start
 import {
   FluentProvider,
   teamsDarkTheme,
   teamsHighContrastTheme,
   teamsLightTheme,
 } from "@fluentui/react-components";
+import { app } from "@microsoft/teams-js";
 import { useTeamsUserCredential } from "@microsoft/teamsfx-react";
 
+import SampleDashboard from "./dashboards/SampleDashboard";
 import { TeamsFxContext } from "./internal/context";
-import SampleDashboard from "./views/dashboards/SampleDashboard";
-import Privacy from "./views/Privacy";
-import TabConfig from "./views/TabConfig";
-import TermsOfUse from "./views/TermsOfUse";
+import Privacy from "./Privacy";
+import TabConfig from "./TabConfig";
+import TermsOfUse from "./TermsOfUse";
 
 /**
  * The main app which handles the initialization and routing
  * of the app.
  */
 export default function App() {
-  const { themeString, teamsUserCredential } = useTeamsUserCredential({
+  const { loading, themeString, teamsUserCredential } = useTeamsUserCredential({
     initiateLoginEndpoint: process.env.REACT_APP_START_LOGIN_PAGE_URL!,
     clientId: process.env.REACT_APP_CLIENT_ID!,
   });
+  useEffect(() => {
+    loading &&
+      app.initialize().then(() => {
+        // Hide the loading indicator.
+        app.notifySuccess();
+      });
+  }, [loading]);
   return (
     <TeamsFxContext.Provider value={{ themeString, teamsUserCredential }}>
       <FluentProvider
@@ -43,15 +46,17 @@ export default function App() {
         }
         className="App"
       >
-        <Router>
-          <Routes>
-            <Route path="/privacy" element={<Privacy />} />
-            <Route path="/termsofuse" element={<TermsOfUse />} />
-            <Route path="/tab" element={<SampleDashboard />} />
-            <Route path="/config" element={<TabConfig />} />
-            <Route path="*" element={<Navigate to={"/tab"} />}></Route>
-          </Routes>
-        </Router>
+        {!loading && (
+          <Router>
+            <Routes>
+              <Route path="/privacy" element={<Privacy />} />
+              <Route path="/termsofuse" element={<TermsOfUse />} />
+              <Route path="/tab" element={<SampleDashboard />} />
+              <Route path="/config" element={<TabConfig />} />
+              <Route path="*" element={<Navigate to={"/tab"} />} />
+            </Routes>
+          </Router>
+        )}
       </FluentProvider>
     </TeamsFxContext.Provider>
   );
