+++ developer-assist-dashboard/src/TabConfig.tsx
