--- developer-assist-dashboard/src/data/ListData.json
@@ -1,17 +1,0 @@
-[
-    {
-        "id": "id1",
-        "title": "Lorem ipsum",
-        "content": "Lorem ipsum dolor sit amet"
-    },
-    {
-        "id": "id2",
-        "title": "Lorem ipsum",
-        "content": "Lorem ipsum dolor sit amet"
-    },
-    {
-        "id": "id3",
-        "title": "Lorem ipsum",
-        "content": "Lorem ipsum dolor sit amet"
-    }
-]
