+++ developer-assist-dashboard/src/internal/login.ts
@@ -1,12 +1,12 @@
 import { TeamsUserCredentialContext } from "./singletonContext";
 
 export async function loginAction(scope: string[]) {
-    try {
-        var credential = TeamsUserCredentialContext.getInstance().getCredential();
-        await credential.login(scope);
-        TeamsUserCredentialContext.getInstance().setCredential(credential);
-    } catch (e) {
-        console.log(e);
-        throw e;
-    }
+  try {
+    var credential = TeamsUserCredentialContext.getInstance().getCredential();
+    await credential.login(scope);
+    TeamsUserCredentialContext.getInstance().setCredential(credential);
+  } catch (e) {
+    console.log(e);
+    throw e;
+  }
 }
